Progressing on processes on mock implementation of geth loaders

+++ cortex-processing/src/main/kotlin/com/hileco/cortex/processing/geth/GethBlock.kt
@@ -1,5 +1,7 @@
 package com.hileco.cortex.processing.geth
 
+import java.math.BigDecimal
+
 class GethBlock(
-        val number: Int
+        val number: BigDecimal
 )
