+++ cortex-processing/src/main/kotlin/com/hileco/cortex/processing/geth/GethBlockchainState.kt
@@ -1,5 +1,7 @@
 package com.hileco.cortex.processing.geth
 
+import java.math.BigDecimal
+
 data class GethBlockchainState(
-        val latestBlock: Int
+        val latestBlock: BigDecimal
 )
